requirements.txt
flask
flask_script
flask_migrate
flask_debugtoolbar
flask_bootstrap
flask_restful
flask_moment
flask_login
flask_wtf
flask_session
flask_assets
flask_principal
flask_compress
flask_htmlmin
flask_mail
flask_admin
sqlalchemy==1.2.0b3
awesome-slugify
htmlmin
jac
pyjwt
pillow
psycopg2
validate_email
celery
redis
